requirements.txt
icoscp==0.1.17
matplotlib==3.8.3
seaborn==0.12.1
pyyaml==6.0

